requirements.txt
pandas>=2.0.0,<3.0.0
tqdm>=4.60.0,<5.0.0
pydantic>=2.0.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
openai>=1.0.0,<2.0.0
